+++ test.py
@@ -1,0 +1,13 @@
+"""Interactive notebook for testing"""
+
+#%% Import packages
+from pathlib import Path
+import numpy as np
+import pandas as pd
+import xlwings as xw
+from utils.config import ROOT_DIR
+
+#%% Create instance of class
+path = Path(r'optimizer_demo.xlsx')
+print(path)
+# %%